+++ destinations/airbyte-faros-destination/src/converters/circleci/models.ts
