+++ destinations/airbyte-faros-destination/src/converters/circleci/pipelines.ts
