--- destinations/airbyte-faros-destination/test/resources/circle-ci/all-streams.log
@@ -1,2 +1,0 @@
-{"record":{"stream":"mytestsource__circle_ci__projects","emitted_at":1650379632259,"data":{"slug":"gh/huongtn/sample-test","organization_name":"huongtn","organization_id":"988b9f2e-fc7c-4a81-b2ac-651c352e01b5","name":"sample-test","id":"19713f32-e5b8-4f7e-9978-ecc3f96b3113","organization_slug":"gh/huongtn","vcs_info":{"vcs_url":"https://github.com/huongtn/sample-test","default_branch":"main","provider":"GitHub"}}},"type":"RECORD"}
-{"record":{"stream":"mytestsource__circle_ci__pipelines","emitted_at":1650379637691,"data":{"id":"2b2a3332-6b7f-47e2-a121-fdc600298175","errors":[],"project_slug":"gh/huongtn/sample-test","updated_at":"2022-04-19T01:31:51.131Z","number":1,"state":"created","created_at":"2022-04-19T01:31:51.131Z","trigger":{"received_at":"2022-04-19T01:31:48.820Z","type":"api","actor":{"login":"huongtn","avatar_url":"https://avatars.githubusercontent.com/u/4100100?v=4"}},"vcs":{"origin_repository_url":"https://github.com/huongtn/sample-test","target_repository_url":"https://github.com/huongtn/sample-test","revision":"4bb2547c19b752c1260fbae2891145faeed63670","provider_name":"GitHub","branch":"circleci-project-setup"},"workflows":[{"pipeline_id":"2b2a3332-6b7f-47e2-a121-fdc600298175","id":"ac658743-d425-4656-b650-60a84cd838aa","name":"say-hello-workflow","project_slug":"gh/huongtn/sample-test","status":"success","started_by":"ab819fe8-314f-413b-80d6-b9007cb47b6d","pipeline_number":1,"created_at":"2022-04-19T03:15:02Z","stopped_at":"2022-04-19T03:15:07Z","jobs":[{"dependencies":[],"job_number":3,"id":"4ce0781d-658e-4155-9215-f26c9bbfb925","started_at":"2022-04-19T03:15:05Z","name":"say-hello","project_slug":"gh/huongtn/sample-test","status":"success","type":"build","stopped_at":"2022-04-19T03:15:07Z"}]},{"pipeline_id":"2b2a3332-6b7f-47e2-a121-fdc600298175","id":"cf43e911-c2c0-4caf-83e3-91e8ec947a81","name":"say-hello-workflow","project_slug":"gh/huongtn/sample-test","status":"success","started_by":"ab819fe8-314f-413b-80d6-b9007cb47b6d","pipeline_number":1,"created_at":"2022-04-19T03:14:50Z","stopped_at":"2022-04-19T03:14:55Z","jobs":[{"dependencies":[],"job_number":2,"id":"12890fdc-f836-4b02-9eab-447115d18075","started_at":"2022-04-19T03:14:52Z","name":"say-hello","project_slug":"gh/huongtn/sample-test","status":"success","type":"build","stopped_at":"2022-04-19T03:14:55Z"}]},{"pipeline_id":"2b2a3332-6b7f-47e2-a121-fdc600298175","id":"851c8211-f38d-414c-8798-0ef9f41e7510","name":"say-hello-workflow","project_slug":"gh/huongtn/sample-test","status":"success","started_by":"ab819fe8-314f-413b-80d6-b9007cb47b6d","pipeline_number":1,"created_at":"2022-04-19T01:31:51Z","stopped_at":"2022-04-19T01:31:56Z","jobs":[{"dependencies":[],"job_number":1,"id":"a1c28fc8-c1ef-4f31-a69c-4a7a3896b0b6","started_at":"2022-04-19T01:31:54Z","name":"say-hello","project_slug":"gh/huongtn/sample-test","status":"success","type":"build","stopped_at":"2022-04-19T01:31:56Z"}]}]}},"type":"RECORD"}
