+++ destinations/airbyte-faros-destination/test/resources/circleci/catalog.json
@@ -2,13 +2,13 @@
   "streams": [
     {
       "stream": {
-        "name": "mytestsource__circle_ci__projects"
+        "name": "mytestsource__circleci__projects"
       },
       "destination_sync_mode": "append"
     },
     {
       "stream": {
-        "name": "mytestsource__circle_ci__pipelines"
+        "name": "mytestsource__circleci__pipelines"
       },
       "destination_sync_mode": "append"
     }
