+++ sources/circleci-source/resources/schemas/pipelines.json
