+++ sources/circleci-source/resources/schemas/projects.json
