+++ sources/circleci-source/resources/spec.json
