+++ sources/circleci-source/test_files/full_configured_catalog.json
