+++ sources/circleci-source/test_files/incremental_configured_catalog.json
