+++ sources/circleci-source/test_files/invalid_config.json
