+++ sources/circleci-source/test_files/pipelines.json
