+++ sources/circleci-source/test_files/pipelines_input.json
