+++ sources/circleci-source/test_files/projects.json
